Re-organize example
