+++ app/over_react_redux/todo_client/example/make_styles/main.dart
@@ -2,7 +2,7 @@ import 'dart:html';
 
 import 'package:over_react/react_dom.dart' as react_dom;
 
-import './mui_components/example_entrypoint.dart';
+import 'example_entrypoint.dart';
 
 main() {
   final container = querySelector('#example-container');
