+++ app/over_react_redux/todo_client/example/with_styles/main.dart
@@ -1,0 +1,13 @@
+import 'dart:html';
+
+import 'package:over_react/react_dom.dart' as react_dom;
+
+import 'example_entrypoint.dart';
+
+main() {
+  final container = querySelector('#example-container');
+
+  final app = ExampleEntrypoint()();
+
+  react_dom.render(app, container);
+}
