+++ app/over_react_redux/todo_client/example/with_styles/with_styles_example.html
@@ -19,20 +19,19 @@
 <head>
     <meta http-equiv="X-UA-Compatible" content="IE=edge, chrome=1">
     <meta charset="utf-8">
-    <title>MUI OverReact Example</title>
+    <title>MUI OverReact With Styles Example</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="apple-mobile-web-app-capable" content="yes">
     <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
 </head>
-    <body>
-        <a href="./make_styles/make_styles_example.html"><h3>Make Styles Examples</h3></a>
-        <a href="./with_styles/with_styles_example.html"><h3>With Styles Examples</h3></a>
+<body>
+<div id="example-container"></div>
 
-        <script src="packages/react/react.js"></script>
-        <script src="packages/react/react_dom.js"></script>
-        <script src="https://unpkg.com/@material-ui/core/umd/material-ui.development.js" crossorigin="anonymous"></script>
+<script src="packages/react/react.js"></script>
+<script src="packages/react/react_dom.js"></script>
+<script src="https://unpkg.com/@material-ui/core/umd/material-ui.development.js" crossorigin="anonymous"></script>
 
-        <!-- Serve compiled output. -->
-        <script defer src="main.dart.js"></script>
-    </body>
+<!-- Serve compiled output. -->
+<script defer src="main.dart.js"></script>
+</body>
 </html>
